requirements.txt
anthropic==0.86.0
beautifulsoup4==4.14.3
Flask==3.1.3
gunicorn==23.0.0
lxml==6.0.2
python-dotenv==1.2.2
python-pptx==1.0.2
requests==2.32.5
